pyproject.toml
[tool.poetry]
name = "odc-python"
version = "0.1.0"
description = "Python API for www.offiziellecharts.de. A minimal toolkit to obtain the top 100 music charts from germany."
authors = ["An Hoang <an.hoang@statworx.com>"]
readme = "README.md"
repository = "https://github.com/AnH0ang/odc-python"
homepage = "https://github.com/AnH0ang/odc-python"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"
pandas = "^1.2.4"
numpy = "^1.20.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b1"
flake9 = "^3.8.3"
mypy = "^0.812"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 100
atomic = true
